PATCH: Link to account profile page in default _user_util_links.erb is broken.
http://jira.projectblacklight.org/jira/browse/CODEBASE-191

The _user_util_links.erb view was generating a link to the logged in
user's profile with user_path(current_user). However, since :user was
mapped as a singleton resource in Blacklight routes, this generates an
incorrect path. The current call is simply user_path.

This patch fixes that, and adds a cucumber feature to check for
existence of a link to user profile page for a logged in user, and that
the link indeed goes to the user profile page.