Remove Facets.paginate class method.  Put logic directly in solr_helper#get_facet_pagination
It's better to have this logic in a method hanging off a Controller (as the
solr_helper mix-in does) than to have it in a global singleton (as a
class method is).  This way in the future this logic can be dependent on
controller-specific parameters/config.

The extra level of indirection having get_facet_pagination call another method
didn't seem useful to me, it just seemed confusing. So I merged it all
together, rather than just moving the singleton paginate method to the
solr_helper.