catalog/facet request parameters (for sort and offset) need to not collide with catalog/index request parameters, so we can pass the catalog/index parameters as limiters (and later pass them back) without confusing things.
The request parameters key names used by by the pager for catalog/facet are now
stored in a class-level hash on Paginator, so we can easily change them in
one place if we change our minds about them.

Also, so the complete list of keys can be obtained in code, to strip out
catalog/facet request params before redirecting back to catalog/index on
a facet click.