Adds the add_facet_limit_from_facet_action helper method used for linking from a facet value in the catalog/facet browse screen.
This is kind of a weird method, in both name and implementation. We already
had the #add_facet_params method, why couldn't we just re-use that? The new
method indeed calls that older method, but does two things it did not:
redirect specifically to the 'index' action instead of just staying with
the current context action. And two, remove the catalog/facet action's
own request parameters, so they don't pollute the URL of the redirected
catalog/index action.

Is there a simpler name for this action? Or should those extra functions
just be made standard in the #add_facet_params method, so this extra
method is not needed? Not sure.