Move individual facet rendering code into helper methods render_facet_value and render_selected_facet_value.
This DRYs up code between the sidebar _facets partial and the catalog/facet expanded facet list.
It also provides useful hooks for over-riding facet item display in a local app or subsidiary plugin,
which is useful to have.

Even though these are helper methods, since they generate URLs known issues with rspec make
it challenging to test them with rspec. We'll have to relie on cucumber.