method delete_or_assign_search_session_params needs to actually start a new hash in session, not merge into previous one.
Merging, after the blacklist-instead-of-whitelist refactor, was causing parts of previous searches end up in the 'current' search hash.

It would be good to have a Feature written to make sure this stuff works right, I messed around with it for a while but couldn't figure out how to get the right feature in there, sorry.