Move the "showing X through Y" message OUT of _constraints partial -- it is not a constraint, putting it there makes things confusing and difficult for customization.  Note that this "showing X through Y" messgae, as well as the "showing X of N" on show page message, are known as "page entry info" in code, not sure why, that's from earlier.
With that info not in _constraints, both index and show view need to then include it seperately. They do it with the #page_entries_info (already existing) and #item_page_entry_info (new with this commit) helper methods.

NOTE WELL: If you have locally customized the index or show viws, and are expecting _constraints to include this display, it no longer does, you may have to add it to your local index and/or show views.