Better active state display for constraint remove button.
The accessible text we put in the link with text-indent:-9999px makes the active state outline all weird. So we remove the active state outline. But for other accessibility (people who tab through links), you really need a distinguishable active state. So we add it back in with a 'border' on :active instead, the border displays less weirdly than the built in outline.