No need to wrap render previous_next_doc in javascript document.write.
The rendering of previous_next_doc on item show was wrapped in javascript that first checked JS history(), and only put the render on the page if there was a history. I think this was an artifact from some previous implementation of next/prev, it is not neccesary to check history() client-side first, the next/prev buttons show up properly with a search, or do not show up if there is no search in session, without it.

Removing this, and it's document.write, removes a layer of confusing JS indirection, and makes this code somewhat easier to follow and simpler. The previous_next_doc partial itself still includes the weird embedded JS for our current next/prev implementation though. I'd really like to entirely change that implementation at some point, but not today.

All tests pass with this change, although I suspect none of the tests target this functionality, since the JS nature of it makes it hard to target with automated tests. I manually tested to my satisfaction.