Remove recently added search box on show page.
I added that in when doing the UX refresh, because it seemed to be consistent and made some CSS easier. However, it doesn't actually work consistently in our somewhat weird search-context-on-show-page system. And it wasn't actually vetted by other committers. So having second thoughts, remove it, now it's more like it was on show page.