Float 'document functions' on results page, instead of YUI-grids.
"document functions" is what code calls the add/remove Booksmarks and Folder buttons. Previously they were seperated from an item title using YUI-grids. This had problems -- it reserved space for the buttons even when they weren't actually there, limiting space for a long title unneccesarily. It also seemed to cause titles to wrap kind of weirdly.

Using an ordinary float here has advantages -- if the floated section is not present, then no space is reserved for it. Even when it is there, a very lengthy title can wrap around the bottom of the 'document function' buttons, which is advantageous. Additionally, HTML becomes somewhat more semantic in this setup, I could eliminate a <div> around the title heading that had no purpose but for YUI-grids to latch onto it.

One possible downside: In order to keep this floated section from messing up the<dl> used for key/value display, I needed to add clear:both to that dl. [That dlas kind of fragile css to make it fairly robust side-by-side dt and dd]. This means that when both buttons are present, and there is a short title, there can be a gap before the <dl> that probably shouldn't be there. However, I think this was true before this change as well; there is room for improvement, but the CSS is tricky. I am considering putting the two action buttons side by side isntead to ameliorate.