Float 'bookmark' and 'folder' buttons NEXT to each other on results page.
I think this is no worse and possibly better than having them top and bottom, but more importantly it eliminates the problem with extra whitespace above the <dl> clearing the floats.

"Folder" is on the right-most margin; since Folder is always there, but Bookmark only when you are logged in, this keeps Folder from moving around when you log in, instead Bookmark simply appears next to it.

Bookmark/Folder were already floated next to each other on show page.