fix double-escaping of Bookmark title
No need to h() something before passing it as a value to hidden_field. Doing so causes double-escape, which made quote marks in a title show up escpaed in bookmarks display. Sorry, couldn't manage to write cucumber feature on this, cucumber didn't want to find '&quot;' or a succesful double-quote in response properly.