rake task respects :replicate_master_url config
In solr.yml, in addition to :url for a given environment, you can now specify a :replicate_master_url. The solrmarc indexing rake task will use :replicate_master_url, if present, as taking precedence for where to index to. I'd add this as a comment example in the new app default solr.yml, but I can't tell where that comes from