add ajaxy checkbox inside existing form
instead of hiding the whole form and putting checkbox/label raw
before it, hide all of form's children and put new checkbox/label
inside existing form. This works just as well, and I think is somewhat
less confusing for the developer and in the DOM. It also means that
CSS positioning on the exiting form still controls the new inputs,
which is less obtrusive.