Add id to email link in folder recognized by ajaxyDialog
code that applies ajaxyDialog recognizes the id and makes the dialog
wider. id was present on individual record, but was not on Folder.

Although i wonder if this should be a callout to a partial anyway for over-riding, or if it should even be a callout to the SAME partial.