SMS carrier choice is sorted alphabetical
Passing a hash to Rails options_for_select, the order of the options is unpredictable, as ruby 1.8 hash is unsorted. Instead, you can pass an array of 2-element-arrays (tuples). Which conveniently ruby Hash#to_a creates. So do a #to_a, and then sort, and then add in the "Please select" prompt FIRST in the list.

Added a spec for the empty/prompt 'please select' option being first, but was unable to figure out how to add a spec for the options actually being sorted using rspec, sorry.