Move legacy :qt passtrhough into add_query_to_solr block.
It's about the search fields, it should go with the other stuff about
the search fields. I kinda want to remove this legacy behavior, but
it's not that big a deal to leave it, and I think Stanford and/or UVa
may still be using it, and there's a spec for it right now and my goal
is not to change any rspec'd behavior in this refactor.