move all handling of user_params[:facets] into #add_facetting_to_solr
was in "whitelist".  Why copy it from user_params to solr_params in one place, only to then modify the solr_params you copied in another?  Do it all together, in the method intended to handle generation of facetting-related params.

Likewise with user_params["facet.fields"], which I think is related to same functionality.

It is unclear WHY BL is doing this stuff, and it doesn't seem to be rspec'd. (If I remove it tests still pass). But trying to leave existing semantics the same in this refactor.

This approach we're taking in general is making it a lot clear how to put code together that is working on the same thing, and eliminate weird intermediate confusing steps. And the rspec tests on solr_search_params help verify that you're not changing expected behavior, as you do it a little at a time. woo.