Spec formerly un-spec'd mapping of incoming "facet.field" and "facets"
Based on what as far as I can tell they were supposed to do, logic was
simplified and normalized a bit too. To use outgoing Solr params direct
and not rely on RSolr among other things.

apparently 'facet.field' and 'facets' in incoming user_params were
synonyms, someone added the same functionality twice not realize it
was already there, because the code was un-understandable. Oh well.
Leep legacy behavior in this refactor.