remove path lookup from solr_marc.rake no longer applicable
rake solr:marc:index theoretically tries to look in several places for a variety of things, including both local app and plugin. However, it's plugin lookup stuff was still based on old version of plugin, was not working. Remove that. Now it pretty much insists on finding stuff in local app, which is not ideal or as originally intended. But making it work in the plugin is tricky, SolrMarc itself is weird with relative paths. Better to just remove the non-working code then to leave non-working code.