respect sleep_after_start param passed in
Was original intention. We were passing in sleep_after_start=>2 (seconds) on our spec/cucumber tasks; was being ignored, it was NOT sleeping after start. Was working anyway. Maybe we don't need to pass in sleep after start? But the jetty_solr_server should respect it if passed in. (2s is already better than the 5s it used to insist on waiting)