Fixing over-ridden will_paginate page_entries_info to not break will_paginate
We over-rode to change the display string to "X-Y of Z", but in the process
broke this method for ordinary will_paginate use with actual ActiveRecords,
at least with recent versions of will_paginate. This made it so you couldn't
use will_paginate ordinarily in a BL app. Should be somewhat better now.

1. Why the heck was our logic looking at controller iVar @response?? Wrong
wrong wrong. Now it just looks at the passed-in argument, doesn't depend
on controller state.

2. WillPaginate::Collection has a method #total_entries, but
RSolr::Ext::Response calls it's same method #total.  To make
our custom page_entries_info work with both, had to do a respond_to?
checking which we've got. Would be better if RSolr::Ext::Response
succesfully duck-typed the WillPaginate::Collection with a #total_entries
instead, but oh well.