fixing MARC encoding issues (warning: possible HACK)
in order to solve issues with index.atom.builder tests, updated
marc gem to 0.4.3, to use marc.bytes[] to gather a substring
rather than marc[] (which is not compatible with 1.9)

to fix problem with templates using raw ASCII-8BIT encoded
values, dropped a fix into render_field_value to always force
encoding to UTF-8 if possible. Otherwise, any non UTF-8 will
cause the template to crash. Not sure if this is a desired
patch, but disabling line 285  will undo it.

Changed a bunch of files with encoding: unknown to UTF-8