make emails text/plain for real
by changing template names to end in .text.erb. Not sure why, but before, with templates just .erb, Rails was putting a <meta> content-type text/html in the email bodies. But I coudn't actually duplicate this in the spec, the spec was NOT doing that, so spec is unchanged. Sorry!

Also, once it's really for real text/plain, Rails3 newline control in text ERB templates is a lot more predictable, took out lots of hyphens from <%- -%> that are no longer neccesary to get newline output we want, they were actually suppressing newlines we didn't want suppressed as is. Again, unable to duplicate this in 'spec', sorry.