another try to make pagination CSS cause display to be how it used to be
the HTML DOM is different, the CSS has to be differnet. Trick every <a> is in a <span> now, didn't used to be, need to keep spacing from double applying, but still apply to current page span, but have hover over a still work right and <a> still be a big target. sigh.