refactored stylesheets to use sass, added better theme support
* refactored most scss to use consistent style
* condensed many nested selectors
* updated compass to 0.12.alpha.2
* abstracted almost all colors into 'variables' document
* assets generator should be idempotent

Now has support for themes:

while requiring 'blacklight/blacklight' will always just load the
default theme, the generator now copies in a file
blacklight_themes/standard, which can be required instead
of blacklight/blacklight to make local changes to colors and grid
size and other common features. All of the variables there are
defaulted elsewhere in the stylesheet, so it should be reliable, even if
we change what variables are needed later.