Remove the Blacklight 'folders' feature
The folders feature is subsumed by bookmarks, by allowing both
logged-in and anonymous users to create bookmarks.

Out-of-the-box, it works with Devise + devise-guests, but should be
adaptable to other systems as well.

Previously, other authentication systems had to provide #current_user,
and now, if they want anonymous bookmarks, will have to also provide
 #current_or_guest_user and #guest_user methods.