Increase click area of facet collapse target and add anchor for accessibility.
This moves the Bootstrap collapse data attribute/classes up the DOM to the panel
heading which increases the click area to the entire panel heading.

This also adds the anchor back into facet heading so that they are in the tab
index and can be expanded/collapsed with the tab and return keys.

Adding a few basic (pending) javascript tests to validate that the facet collapse is happening.