No need to check for Turbolinks defined.
If there's no Turbolinks, we registered an event handler
for an event that will never be hired -- this should not harm
anything.

But the check introduces possible order-of-load errors, where
blacklight core accidentally ends up loaded before turbolinks,
so doesn't set up the event handler, but then turbolinks loads later,
so the event handler was needed, but didn't load. ran into this
with an engine that inserted it's 'require x.js' before turbolinks.
Best to avoid order-of-load dependencies if we can; we can.