Ignore empty f[facet][] query param
Previously this would result in a limit being applied
for a facet with an empty string value. I don't think this
is a real use case (it would be somewhat tricky to index
values in Solr this way even if you wanted to), and I
don't think this was intentional.

Seems better to ignore empty facet limit query params
like that, have them be no-ops that do not effect
the Solr query or the visible constraints. That's
what this does.

empty facet limit query params of course shouldn't happen
at all, and are probably mistakes, but could come from:
* a bug in some part of BL or plugin
* a bug in some external "deep link generating" tool
* user error manipulating the URL manually

I suggest it's better to have such a mistake be a no-op, then
to have it result in odd and disconcerting behavior as previous.