Fix Next and Previous Links still clickable when disabled.
While they are greyed out and show a disabled icon, the next and
previous pagination navigation is still clickable and will lead to
a weird state in the case of "next". This fixes it in a very easy
to read manner. There may be a more compact way of making this
change.

To duplicate the symptoms this fixes (tested in Chrome):

* Go to http://demo.projectblacklight.org/ and do a blank search.

* Click "<< Previous" at the very bottom of the page for error #1.
This is minor as it simply refreshes the page.

* Click "988" at the bottom of the page (the last page).

* Click the greyed-out "Next >>" at the bottom. This will lead to a
minor but more undesired state.