Add separate setters to search_builder
Fixes #1462

If we wanted to drop/limit the builder pattern we would insert
deprecation warnings in those methods to be converted to pure getters.
If we want to keep the builder pattern, then we let it ride with the new explicit setters.

Updated calls where the returned `self` wasn't being used and the basic setter pattern is more idiomatic.

Added tests to make sure builder-style w/ arg calls underlying setter method.