Normalize the bookmark_controller.delete method.
Currently the `bookmark_controller.create` method allows the user to
pass an array of bookmarks via the parameters with a format of
`[{ id: foo, document_type: bar }]`.

This is useful because theoretically we can bookmark records that are
not of type SolrDocument.  Likewise the bookmark model does not make any
assumptions about document_type other than providing a default if one is
not given.

Contrary to these methods, the `bookmark_controller.delete` is not
implemented in a consistent fashion to its counterpart.

`bookmark_controller.delete` works on only one bookmark at a time and
its document type can only be derived from `blacklight_config`. This is
unfortunate because `blacklight_config` is copied from the
`CatalogController` and the document type is already set to be
`SolrDocument`.

This commit updates `bookmark_controller.delete` to work in
a complimentary manner to `bookmark_controller.create`.

The use case of this change is to allow for being able to bookmark non
SolrDocuments in an environment where there are multiple backends not
just of type Solr.

* This change also fixes bookmark creation via bookmarks params not working in Rails when strong parameters is enabled.