Add merge_facet method to include in solr response.
This commit adds the method `merge_facet` which comes in handy when
creating a federated facets that merges results from multiple sources.

This seems to be a missing method since it is already possible to add
a :url_method for configuring facet fields, which suggest it's already
possible to potentially federate a facet field.