Match bootstrap4 accordion example, avoids 'jankiness' on facet open/close
Some had noticed that Blacklight 7 seemed to have "jankiness" in the slide up/down animation for opening/closing facet areas.

From past experience, I guessed this may have to do with "margin" or "padding" on the div that is the the target for actual collapse/expand. Debugging, I confirmed that removing "padding" removed jankiness for me. But of course we want padding. Did that suggest we need a separate 'wrapper' div?

Consulting the bootstrap4 example for using a bootstrap4 "card" as the target of an expand/contract, at: https://getbootstrap.com/docs/4.1/components/collapse/

I noticed the canonical example DID have a separate div -- there was a wrapper div that was the collapse target, which contained a separate div which was the actual 'class="card-body"' As opposed to Blacklight before this PR, which used the card-body div itself as the collapse target.

Altering Blacklight's DOM to use two separate divs as in the canonical Bootstrap4 example -- for me, eliminated the observed "jankiness".

In order for blacklight_range_limit to keep working, I needed to move the classes 'panel-collapse facet-content' to the collapse-target-wrapper div as well. Not totally sure what's going on there, but blacklight_range_limit in my demo app worked once I moved these classes; did not without moving them.