Add docker-compose configuration for local development.
This commit  adds a docker-compose.yml and Docker definitions to set up
a local blacklight development.

In order to bring up an instance of the project locally the developer will
need to install docker and then run:

`docker-compose up`

from the root of the project.

The solr instance will be available at http://localhost:8983
The blacklight app instance will be available at http://localhost:3000

Note that the current definition continues to use the enginecart gem for
setting up the instance as this makes sense for development of a rails
engine application.

Also note that this commit does not add docker to the internal project
that gets generated.

This commit also removes an unused Vagrant file.