Remove unneeded auto-loader config
I believe this config should be unnecessary, and possibly could cause a problem with the new Rails6 zeitwerk auto-loader. May not cause a bug, but the code that isn't even in your codebase is the code that can't possibly cause problems!

https://guides.rubyonrails.org/upgrading_ruby_on_rails.html#autoloading

app/presenters, along with all sub-dirs of `./app`, should be automatically put in load path by Rails.
Similary, we are usingly the ./concerns subdirs in a standard Rails way, which Rails should handle by default.

I suspect this config is left over from long ago versions of Rails before it was fully default behavior.

Note `config.root` in this config is Blacklight engine root, not app root.