Adds autocomplete="off" when suggest/autocomplete feature is enabled
The current version of the twitter typeahead library specified in Blacklight could pose issues to adopters. While this is unfortunate adopters don't have to include this in their JavaScript and in fact can include a replacement. That replacement may or may not automatically provide this, so as a convenience for Blacklight adopters we should just go ahead and this markup.