Do not mutate search_state when mutating search_state.to_h
`Blacklight::SearchState#to_h` needs to return a dupe of it's internal
state @params not a reference to it.  Otherwise whatever mutation you do
to the reference gets reflected in the Blacklight::SearchState instance.

It's unexpected behavior for the `to_h` method to expose the
`Blacklight::SearchState` instance's copy of `@params` and it can lead
to very confusing bugs.