remove four-year-old deprecated -image scss code
I'm not sure what this code is doing or intending to do. I think it is actually buggy code that only triggers in some cases. I don't think it's actually any backwards incompat to remove it, I don't think it was doing anything useful. But I'm not totally positive.

Commit that added this code originally at ebc1a7e

This is one possible closes #2414 see analysis there.

Review and a second opinion is welcome.