Extract components for the headers
This allows us to override the header component, while reusing existing components. For example, at Stanford we would like to insert a hidden third navbar that has a feedback form between the top bar and the search navbar.  This will allow us to override the header component to do that.