Release 7.x backports (#3071)
* Allow the application link in the header to be configured

* Update the route to the autocomplete suggestion endpoint

* Extract components for the headers

This allows us to override the header component, while reusing existing components. For example, at Stanford we would like to insert a hidden third navbar that has a feedback form between the top bar and the search navbar.  This will allow us to override the header component to do that.

* Use container_classes from helpers rather than views

This allows consuming app to override the helper.
Fixes #2935

* Extract a logo_link method on the TopNavbarComponent

This allows you specify the logo text on the component

* Log a warning when the configuration specifies a partial that isn't resolvable (#3026)

* Allow the metadata_component to be configured

This makes it easier to swap out the default component without having to override the _show_main_content.html.erb.  Additionally this provides consistency with how you can set the thubmnail_component

* Allow the title_component to be configured

* Update description in package.json

* ViewComponent 3 changes collection counter from 1-indexed to 0-indexed

* Explicitly set the document collection parameter to maintain API compatibility with subclasses

* Update documentation about subclassing the DocumentComponent

* Add reasoning about why a counter variable must be provided.

* Allow document_index_view_type outside a session context

* Add parameters to metadata components that change the tag, classes, and attributes

* use ViewComponent 3 in one of the CI matrix runs

* Don't populate the q parameter unless a q parameter was provided

* Partial backport of #2698

* Make metadata and thumbnail components backwards compatible with 7.x

---------

Co-authored-by: Justin Coyne <jcoyne@justincoyne.com>
Co-authored-by: Benjamin Armintor <armintor@gmail.com>