Translate negative facets into fq to exclude docs with a given facet value
This commit is a rough first hack at adding the ability to filter out docs from a result set that match a facet value. For instance, all docs with `language_ssim:Tibetan` excluding those with `language_ssim:English`. None of the UI work to support this is included in this commit, which is more to get familiar (once again) with Solr and Blacklight. There is probably a better way to handle this BL query args-to-Solr query translation than injecting `{ exclude: true }` into facet values and sniffing for it. Pushing this commit up for early feedback to come up with a solid approach before tackling UI work or tests.