configure zeitwerk to be okay with our generators in local generated app ./lib/generators
Not sure why we are generating a generator INTO the local test app. But rather than figure this out and change it, we want to make zeitwerk okay with it.

We are telling zeitwerk to ignore the WHOLE RAILS_ROOT/lib/generators directory, just to deal with the one test_app_generator.rb file we are putting there. But hopefully that is okay.