Ensure the matching version of the assets package is installed
Because new npm package versions can be published without setting them to "latest", doing a `yarn add blacklight-frontend`  with no version specifier currently results in retrieving v7.38.0, even though versions for v8.x are published.

This ensures that we get the npm asset version that matches Blacklight's version by asking for it explicitly, even if it has not been marked `latest`. If there is a mismatch between npm package versions and Blacklight gem versions, we'll get an error.