Don't include blacklight js twice when using Sprockets
As @mamrey noted on #3263, this can lead to a user being unable to add bookmarks,
since the relevant eventlistener is added twice, so every time the user clicks to
add a bookmark, the toggle function is run twice, essentially doing nothing.

This bug (and fix) only affect Sprockets -- When using Importmap or Propshaft, this
branch of the view isn't even reached.