Use a labeled <aside> for show tools sidebar
The show page sidebar is currently displayed
inside a <section>; this changes it to use the <aside> element,
which is arguably more appropriate for a sidebar.

The search sidebar has an aria-label already set, but the show
sidebar did not, which also meant that it had no accessible name
despite having a landmark role.

Thus this also adds a configurable aria-label for the show page sidebar
similar to the one for the search sidebar.