Add additional accessible text to linked faceted searches
This updates the LinkToFacet rendering step to add an additional
visually hidden element when creating a link to a faceted search
that indicates the facet that will be used in the search.

This helps give additional context to screen reader users when
link_to_facet is used, especially when the page may include
links that go to searches for the same value in different fields.

Fixes #3427