Ensure that facet suggestions are relevant to the user's query
Before this commit, the facet suggest javascript sent HTTP
requests that included only the name of the facet and the
fragment that the user entered to filter down the facets.

This commit also sends any URL params that were in the user's
original search, so that the facet suggestions are relevant
to the search, rather than being just the most popular facet
values from the entire corpus.

Closes #3494