Un-pend a test for catalog URLs with slashes in the document ID
Ref #911

Solr documents with a slash in the ID seem to route correctly
without additional changes now. I tested by adding a document to
the index and then checking its catalog URL; Rails's link_to
seems to automatically generate a link with an encoded slash.

Unfortunately this does not apply to a period, so full URLs won't
work.