Restore scrolling and execute other modal hide behavior on modal dismiss by "escape" key
Blacklight modals use bootstrap modals, but put them in a HTML `dialog` element -- which is neat, but is not what bootstrap expects.

You can dismiss a modal-in-<dialog> with an escape key -- if this is done, the Bootstrap 'hide' function does not get invoked, it's purely a user-agent action to end the dialog.

This catches event for that to restore BL hide func on escape-to-close, including restoring window scrolling, and custom event.