Switch facet suggest request to have query in URL query params, instead of path, with proper escaping
It was failing prior on queries including `.` or `/`.  It may have been possible to fix this with proper escaping and leave it in path component of URL -- but it's much more straightoforward to put it in query component after the `?`, and let existing escaping happening take care of it.   This is a URL only used by JS fetch that shouldn't ordinarily be seen by the user, so aesthetics prob not a concern.

The Blacklight CI sample data has a `topic_facet` value with a period in it (`Iran. Vizārat-i Kishvar`), and my own local data does too. I haven't seen any with `/`, but of course it's possible and should work.